+++ tests/Klein/Tests/Fixtures/routes/random.php
@@ -1,0 +1,15 @@
+<?php
+
+$this->respond(
+    path: '/?',
+    callback: function ($request, $response, $app) {
+        echo 'yup';
+    }
+);
+
+$this->respond(
+    path: '/testing/?',
+    callback: function ($request, $response, $app) {
+        echo 'yup';
+    }
+);